router for v1 alice-bob endpoints
