+++ src/index.ts
@@ -20,23 +20,17 @@ import { getNotifications } from './notifications/utils/get-notifications.util';
 import { getParsedContent } from './notifications/utils/get-parsed-content.util';
 import { getPlatforms } from './notifications/utils/get-platforms.util';
 import { redisClient } from './redis';
+import { aliceBobRouter } from './routers/alice-bob';
 import { evmRouter } from './routers/evm';
 import { adRulesRouter } from './routers/slise-ad-rules';
 import { templeWalletAdsRouter } from './routers/temple-wallet-ads';
 import { getTkeyStats } from './tkey-stats';
 import { getABData } from './utils/ab-test';
-import { cancelAliceBobOrder } from './utils/alice-bob/cancel-alice-bob-order';
-import { createAliceBobOrder } from './utils/alice-bob/create-alice-bob-order';
-import { estimateAliceBobOutput } from './utils/alice-bob/estimate-alice-bob-output';
-import { getAliceBobEstimationPayload } from './utils/alice-bob/get-alice-bob-estimation-payload';
-import { getAliceBobOrderInfo } from './utils/alice-bob/get-alice-bob-order-info';
-import { getAliceBobPairInfo } from './utils/alice-bob/get-alice-bob-pair-info';
-import { getAliceBobPairsInfo } from './utils/alice-bob/get-alice-bob-pairs-info';
 import { btcExchangeRateProvider, tezExchangeRateProvider } from './utils/coingecko';
 import { CodedError } from './utils/errors';
 import { exolixNetworksMap } from './utils/exolix-networks-map';
 import { coinGeckoTokens } from './utils/gecko-tokens';
-import { getExternalApiErrorPayload, isDefined, isNonEmptyString } from './utils/helpers';
+import { isDefined, isNonEmptyString } from './utils/helpers';
 import logger from './utils/logger';
 import { getSignedMoonPayUrl } from './utils/moonpay/get-signed-moonpay-url';
 import { getSigningNonce } from './utils/signing-nonce';
@@ -219,93 +213,6 @@ app.get('/api/moonpay-sign', async (_req, res) => {
   }
 });
 
-app.post('/api/alice-bob/create-order', async (_req, res) => {
-  const { isWithdraw, amount, from, to, userId, walletAddress, cardNumber } = _req.query;
-
-  try {
-    const payload = {
-      ...getAliceBobEstimationPayload(isWithdraw, from, to, amount),
-      userId: String(userId),
-      toPaymentDetails: isDefined(cardNumber) ? String(cardNumber) : String(walletAddress),
-      redirectUrl: 'https://templewallet.com/mobile'
-    };
-
-    const orderInfo = await createAliceBobOrder(payload);
-
-    res.status(200).send({ orderInfo });
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send(data);
-  }
-});
-
-app.post('/api/alice-bob/cancel-order', async (_req, res) => {
-  const { orderId } = _req.query;
-
-  try {
-    await cancelAliceBobOrder({ id: String(orderId) });
-
-    res.status(200);
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send(data);
-  }
-});
-
-app.get('/api/alice-bob/get-pair-info', async (_req, res) => {
-  const { isWithdraw } = _req.query;
-
-  try {
-    const pairInfo = await getAliceBobPairInfo(isWithdraw === 'true');
-
-    res.status(200).send({ pairInfo });
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send(data);
-  }
-});
-
-app.get('/api/alice-bob/get-pairs-info', async (_req, res) => {
-  const { isWithdraw } = _req.query;
-
-  try {
-    const pairsInfo = await getAliceBobPairsInfo(isWithdraw === 'true');
-
-    res.status(200).send({ pairsInfo });
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send(data);
-  }
-});
-
-app.get('/api/alice-bob/check-order', async (_req, res) => {
-  const { orderId } = _req.query;
-
-  try {
-    const orderInfo = await getAliceBobOrderInfo(String(orderId));
-
-    res.status(200).send({ orderInfo });
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send({ error: data });
-  }
-});
-
-app.post('/api/alice-bob/estimate-amount', async (_req, res) => {
-  const { isWithdraw, amount, from, to } = _req.query;
-
-  try {
-    const payload = getAliceBobEstimationPayload(isWithdraw, from, to, amount);
-
-    const outputAmount = await estimateAliceBobOutput(payload);
-
-    res.status(200).send({ outputAmount });
-  } catch (error) {
-    const { status, data } = getExternalApiErrorPayload(error);
-    res.status(status).send({ error: data });
-  }
-});
-
 app.get('/api/mobile-check', async (_req, res) => {
   const platform = _req.query.platform;
   const appCheckToken = _req.query.appCheckToken;
@@ -350,6 +257,8 @@ app.use('/api/slise-ad-rules', adRulesRouter);
 app.use('/api/evm', evmRouter);
 
 app.use('/api/temple-wallet-ads', templeWalletAdsRouter);
+
+app.use('/api/alice-bob', aliceBobRouter);
 
 app.post('/api/magic-square-quest/start', async (req, res) => {
   try {
