+++ src/routers/alice-bob/utils/cancel-alice-bob-order.ts
@@ -1,5 +1,5 @@
-import { AliceBobCancelOrderPayload } from '../../interfaces/alice-bob.interfaces';
-import { aliceBobApi } from '../api.sevice';
+import { aliceBobApi } from '../api';
+import { AliceBobCancelOrderPayload } from '../interfaces';
 
 import { getAliceBobRequestHeaders } from './get-alice-bob-request-headers';
 import { getAliceBobSignature } from './get-alice-bob-signature';
