+++ src/routers/alice-bob/utils/create-alice-bob-order.ts
@@ -1,5 +1,5 @@
-import { AliceBobCreateOrderPayload, aliceBobOrder } from '../../interfaces/alice-bob.interfaces';
-import { aliceBobApi } from '../api.sevice';
+import { aliceBobApi } from '../api';
+import { AliceBobCreateOrderPayload, aliceBobOrder } from '../interfaces';
 
 import { getAliceBobRequestHeaders } from './get-alice-bob-request-headers';
 import { getAliceBobSignature } from './get-alice-bob-signature';
