+++ src/routers/alice-bob/utils/estimate-alice-bob-output.ts
@@ -1,5 +1,5 @@
-import { AliceBobEstimateAmountPayload } from '../../interfaces/alice-bob.interfaces';
-import { aliceBobApi } from '../api.sevice';
+import { aliceBobApi } from '../api';
+import { AliceBobEstimateAmountPayload } from '../interfaces';
 
 import { getAliceBobRequestHeaders } from './get-alice-bob-request-headers';
 import { getAliceBobSignature } from './get-alice-bob-signature';
