+++ src/routers/alice-bob/utils/get-alice-bob-order-info.ts
@@ -1,5 +1,5 @@
-import { aliceBobOrder } from '../../interfaces/alice-bob.interfaces';
-import { aliceBobApi } from '../api.sevice';
+import { aliceBobApi } from '../api';
+import { aliceBobOrder } from '../interfaces';
 
 import { getAliceBobRequestHeaders } from './get-alice-bob-request-headers';
 import { getAliceBobSignature } from './get-alice-bob-signature';
