+++ src/routers/alice-bob/utils/get-alice-bob-request-headers.ts
@@ -1,4 +1,4 @@
-import { EnvVars } from '../../config';
+import { EnvVars } from '../../../config';
 
 export const getAliceBobRequestHeaders = (signature: string, now: number) => ({
   'public-key': EnvVars.ALICE_BOB_PUBLIC_KEY,
