+++ src/routers/alice-bob/utils/get-alice-bob-signature.ts
@@ -1,7 +1,7 @@
 import crypto from 'crypto';
 
-import { EnvVars } from '../../config';
-import { AliceBobPayload } from '../../interfaces/alice-bob.interfaces';
+import { EnvVars } from '../../../config';
+import { AliceBobPayload } from '../interfaces';
 
 export const getAliceBobSignature = (payload?: AliceBobPayload) => {
   const now = Date.now();
